+++ packages/forma-36-react-components/src/components/Grid/GridItem/index.ts
@@ -1,0 +1,3 @@
+export * from './GridItem';
+export { default } from './GridItem';
+