+++ packages/components/forma-36-react-timepicker/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.72](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-timepicker@0.2.0-alpha.71...@contentful/forma-36-react-timepicker@0.2.0-alpha.72) (2020-04-06)
+
+**Note:** Version bump only for package @contentful/forma-36-react-timepicker
+
+
+
+
+
+# [0.2.0-alpha.71](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-timepicker@0.2.0-alpha.70...@contentful/forma-36-react-timepicker@0.2.0-alpha.71) (2020-04-06)
+
+**Note:** Version bump only for package @contentful/forma-36-react-timepicker
+
+
+
+
+
 # [0.2.0-alpha.70](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-timepicker@0.2.0-alpha.69...@contentful/forma-36-react-timepicker@0.2.0-alpha.70) (2020-04-06)
 
 **Note:** Version bump only for package @contentful/forma-36-react-timepicker